Refactoring csv functions Part 1
